move `deleteArchivedRecords` to utils
